% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe_set1.R
\docType{data}
\name{anscombe_set1}
\alias{anscombe_set1}
\title{Anscombe's quartet, set 1}
\format{A data frame with 11 rows and 2 variables:
\describe{
  \item{x1}{the x's}
  \item{y1}{the y's}
  ...
}}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/RHOME/library/datasets/html/anscombe.html}
}
\usage{
anscombe_set1
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
From the datasets package
}
\keyword{datasets}
