% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe_tidy.R
\docType{data}
\name{anscombe_tidy}
\alias{anscombe_tidy}
\title{Anscombe's quartet, tidied}
\format{A data frame with 44 rows and 4 variables:
\describe{
  \item{observation}{which point}
  \item{set}{one, two, three, or four as roman numerals}
  \item{x1}{the x's}
  \item{y1}{the y's}
}}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/RHOME/library/datasets/html/anscombe.html}
}
\usage{
anscombe_tidy
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
From the datasets package
}
\keyword{datasets}
