% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeoff_ratings.R
\docType{data}
\name{bakeoff_ratings}
\alias{bakeoff_ratings}
\title{Ratings for "The Great British Bake-Off"}
\format{A data frame with 74 rows and 10 variables:
\describe{
  \item{series}{an integer denoting UK series (1-8)}
  \item{episode}{an integer denoting episode number within a series}
  ...
}}
\source{
\url{https://github.com/apreshill/bakeoff}
}
\usage{
bakeoff_ratings
}
\description{
Data scraped from Wikipedia
}
\keyword{datasets}
