% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarrraw-package.R
\docType{package}
\name{datarrraw-package}
\alias{datarrraw}
\alias{datarrraw-package}
\title{datarrraw: Making a Package with Data Inside}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Desirée De Leon \email{desiree@rstudio.com} (https://orcid.org/0000-0003-2217-9067)

}
\keyword{internal}
